USFS AirFire PM2.5 timeseries scales

Add USFS AirFire-style x-axis and y-axis scales suitable for a timeseries plot showing PM2.5 data.

custom_pm25TimeseriesScales(
  monitor = NULL,
  startdate = NULL,
  enddate = NULL,
  ylim = NULL,
  timezone = NULL,
  xlab = NULL,
  yexp = c(0.05, 0),
  xexp = c(0, 0.05),
  ...
)

Arguments

monitor

A mts_monitor object.Should match the default dataset of the plot.

startdate

Desired start date (integer or character in ymd format or POSIXct).

enddate

Desired end date (integer or character in ymd format or POSIXct).

ylim

custom y-axis limits. This function will apply a default limit depending on the data.

timezone

Timezone for x-axis scale. If NULL and only one timezone present in monitor, that timezone will be used. If NULL and multiple timezones present, the default is UTC.

xlab

Custom xlab. If NULL a default xlab will be generated.

yexp

Vector of range expansion constants used to add some padding around the data on the y-axis, to ensure that they are placed some distance away from the axes.

xexp

Vector of range expansion constants used to add some padding around the data on the x-axis, to ensure that they are placed some distance away from the axes.

...

Additional arguments passed on to custom_datetimeScale.